requirement.txt
System Requirements
Python 3.7
NVIDIA GPU
CUDA 11.0 and cuDNN 7.6.5

libs of python:
numpy
torch
torchvision
Pillow
matplotlib
skimage
opencv-python



